requirements.txt
flask
werkzeug
sqlite3
flask-discord
python-dotenv
requests